Move interfaces to Core module
