+++ core/build.gradle
@@ -9,9 +9,10 @@ java {
 }
 
 dependencies {
-
     // Koin: Dependency Injection
     implementation "io.insert-koin:koin-core:$koin_version"
+    // Coroutines
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4'
 
     testImplementation 'junit:junit:4.13.2'
 }
