+++ data/src/main/java/com/example/data/di/DataSourceModule.kt
@@ -4,8 +4,8 @@ import com.example.data.datasource.IRemotePodcastDataSource
 import com.example.data.datasource.LocalEpisodeDataSource
 import com.example.data.datasource.LocalPodcastDataSource
 import com.example.data.datasource.RemotePodcastDataSource
-import com.example.podcast_details_domain.data_interfaces.datasource.ILocalEpisodeDataSource
-import com.example.podcast_details_domain.data_interfaces.datasource.ILocalPodcastDataSource
+import com.example.core.data_interfaces.datasource.ILocalEpisodeDataSource
+import com.example.core.data_interfaces.datasource.ILocalPodcastDataSource
 import org.koin.dsl.module
 
 val dataSourceModule = module {
