+++ data/src/main/java/com/example/data/di/RepositoriesModule.kt
@@ -2,8 +2,8 @@ package com.example.data.di
 
 import com.example.data.repositories.EpisodeRepository
 import com.example.data.repositories.PodcastRepository
-import com.example.podcast_details_domain.data_interfaces.repositories.IEpisodeRepository
-import com.example.podcast_details_domain.data_interfaces.repositories.IPodcastRepository
+import com.example.core.data_interfaces.repositories.IEpisodeRepository
+import com.example.core.data_interfaces.repositories.IPodcastRepository
 import org.koin.dsl.module
 
 val repositoryModule = module {
