+++ data/src/main/java/com/example/data/extensions/PodcastExtensions.kt
@@ -3,7 +3,7 @@ package com.example.data.extensions
 import com.example.core.type.Genre
 import com.example.data.PodcastDetailsQuery
 import com.example.data.models.LocalPodcastDetails
-import com.example.podcast_details_domain.models.PodcastDetails
+import com.example.core.models.podcastDetails.PodcastDetails
 
 fun PodcastDetailsQuery.GetPodcastSeries.toDomainPodcast(): PodcastDetails {
     return PodcastDetails(
