+++ podcast-details-ui/src/main/java/com/example/podcast_details_ui/screens/podcastDetails/PodcastDetailsViewModel.kt
@@ -3,7 +3,6 @@ package com.example.podcast_details_ui.screens.podcastDetails
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.core.models.ApiResponse
-import com.example.podcast_details_domain.models.PodcastDetails
 import com.example.podcast_details_domain.useCases.FetchPodcastDetailsUseCase
 import com.example.podcast_details_domain.useCases.GetPodcastDetailsUseCase
 import kotlinx.coroutines.flow.*
