+++ podcast-details-ui/src/main/java/com/example/podcast_details_ui/screens/podcastDetails/components/PodcastDetailsContent.kt
@@ -16,7 +16,7 @@ import com.example.core_ui.components.fillLazyColumn
 import com.example.core_ui.theme.PodcastAppTheme
 import com.example.podcast_details_domain.mocks.mockPodcast
 import com.example.core.models.Episode
-import com.example.podcast_details_domain.models.PodcastDetails
+import com.example.core.models.podcastDetails.PodcastDetails
 
 
 @Composable
